Move the RefactoringChecker to its own file
